NexCar - Final Project/WEBAPP + API + DB/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
paho-mqtt
marshmallow
marshmallow-sqlalchemy
requests
numpy
